Hash update is now working again
The context of the execution of the updateLocation function is living INSIDE the iFrame, and because of that the iframes object is not available ( set to undefined ).
Therefore the usage of the event object is highly suggested. Wrap it into an array to keep the code flowing as the function expects an iterable object.